AssignedLicenses.ps1: bugfix

+++ AzureAD/AssignedLicenses.ps1
@@ -54,17 +54,19 @@ function OutlookKiller {
 $dbfile = "C:\Users\$env:USERNAME\AppData\Local\PatrolDB.csv.AES"
 $dbfile_unlocked = "C:\Users\$env:USERNAME\AppData\Local\PatrolDB.csv"
 
-# reading current key
-$adialog = FormBase -w 400 -h 230 -text "UNLOCK DB"
-RadioButton -form $adialog -checked $true -x 20 -y 20 -w 500 -h 30 -text "Enter the key for accessing to DB file" | Out-Null
-$currentkey = TxtBox -form $adialog -x 20 -y 50 -w 300 -h 30 -text ''
-$cleanDB = RadioButton -form $adialog -checked $false -x 20 -y 80 -w 500 -h 30 -text "Clean existing DB file"
-OKButton -form $adialog -x 100 -y 130 -text "Ok" | Out-Null
-$result = $adialog.ShowDialog()
-if ($cleanDB.Checked -eq $true) {
-    $answ = [System.Windows.MessageBox]::Show("Really delete DB file?",'DELETE','YesNo','Warning')
-    if ($answ -eq "Yes") {    
-        Remove-Item -Path $dbfile
+if (Test-Path $dbfile -PathType Leaf) {
+    # reading current key
+    $adialog = FormBase -w 400 -h 230 -text "UNLOCK DB"
+    RadioButton -form $adialog -checked $true -x 20 -y 20 -w 500 -h 30 -text "Enter the key for accessing to DB file" | Out-Null
+    $currentkey = TxtBox -form $adialog -x 20 -y 50 -w 300 -h 30 -text ''
+    $cleanDB = RadioButton -form $adialog -checked $false -x 20 -y 80 -w 500 -h 30 -text "Clean existing DB file"
+    OKButton -form $adialog -x 100 -y 130 -text "Ok" | Out-Null
+    $result = $adialog.ShowDialog()
+    if ($cleanDB.Checked -eq $true) {
+        $answ = [System.Windows.MessageBox]::Show("Really delete DB file?",'DELETE','YesNo','Warning')
+        if ($answ -eq "Yes") {    
+            Remove-Item -Path $dbfile
+        }
     }
 }
